Add files via upload
This pull request focuses on utilizing the "listingAndReviews" dataset from the "sample_airbnb" collection. It involves the implementation of MongoDB queries, incorporating logical operators such as $and, $or, $nor, $not, as well as comparison operators like $eq, $gt, $gte, $lt, and $lte. Additionally, the pull request includes Python code for establishing a connection to a MongoDB server using the PyMongo library